cam restart

+++ experiments/sent_clf/baselines/roberta_pytorch_finetune.py
@@ -234,17 +234,16 @@ if __name__ == '__main__':
                             logger.info(f"***** (Embeds and) Test - Fold {fold_name} *****")
                             logger.info(f"  Details: {best_val_res}")
 
-                            for EMB_TYPE in ['poolbert', 'avbert']:
-                                feat_fp = os.path.join(FEAT_DIR, f"all_features.pkl")
-                                all_ids, all_batches, all_labels = load_features(feat_fp, batch_size=1, sampler=SAMPLER)
-
-                                tmp = [i for i in all_ids if '5nyt1' in i]
-                                embs = inferencer.predict(model, all_batches, return_embeddings=True, emb_type=EMB_TYPE)
-                                basil_w_BERT = pd.DataFrame(index=all_ids)
-                                basil_w_BERT[EMB_TYPE] = embs
-                                emb_name = f'{name}_basil_w_{EMB_TYPE}'
-                                basil_w_BERT.to_csv(f'data/{emb_name}.csv')
-                                logger.info(f'Written embs ({len(embs)},{len(embs[0])}) to data/{emb_name}.csv')
+                            for EMB_TYPE in ['poolbert', 'avbert', 'unpoolbert']:
+                                emb_fp = f'data/{name}_basil_w_{EMB_TYPE}'
+                                if not os.path.exists(emb_fp):
+                                    feat_fp = os.path.join(FEAT_DIR, f"all_features.pkl")
+                                    all_ids, all_batches, all_labels = load_features(feat_fp, batch_size=1, sampler=SAMPLER)
+                                    embs = inferencer.predict(model, all_batches, return_embeddings=True, emb_type=EMB_TYPE)
+                                    basil_w_BERT = pd.DataFrame(index=all_ids)
+                                    basil_w_BERT[EMB_TYPE] = embs
+                                    basil_w_BERT.to_csv(f'data/{emb_name}.csv')
+                                    logger.info(f'Written embs ({len(embs)},{len(embs[0])}) to data/{emb_name}.csv')
 
                             # eval on test
